feat(PN-14054): refactoring package
