+++ src/main/java/it/pagopa/pn/portfat/PortFatApplication.java
@@ -1,4 +1,4 @@
-package it.pagopa.pn.template;
+package it.pagopa.pn.portfat;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
